feat: improve rendering. Make positions more precise. Indicate truck direction of travel, display queues better. Add more trucks and make dump take longer to properly test dump queue.

+++ TruckSim/Entities/Entity.h
@@ -22,7 +22,7 @@ public:
         return position;
     }
     
-    void SetPosition(int x, int y)
+    void SetPosition(float x, float y)
     {
         position = Position{x,y};
     }
